api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
slowapi==0.1.9
pyyaml==6.0.1
python-jose[cryptography]==3.3.0

